one more swiper width fix for windows + card carousel added as a component
