+++ src/example-pages/components/promo/VCOD.vue
@@ -24,7 +24,7 @@
           <div class="rt-col-2 td-d-none"></div>
         </div>
         <div class="row">
-          <div class="custom-carousel">
+          <rt-card-carousel>
             <rt-card background-color-type="dark-slate" :background-blur="5" :is-white-color="true" class="carousel-card carousel-card--active">
               <template slot="content">
                 <div class="rt-space-top15 rt-space-bottom">
@@ -186,32 +186,32 @@
                   множества проектов федерального масштаба</p>
               </template>
             </rt-card>
-            <div class="custom-carousel__nav">
-              <div class="custom-carousel__nav-item custom-carousel__left" @click="previousSlide">
-                <svg width="12px" height="7px" viewBox="0 0 12 7" version="1.1" xmlns="http://www.w3.org/2000/svg" class="custom-carousel__left-arrow">
-                  <g id="Page-1" stroke-width="1">
-                    <g transform="translate(-132.000000, -2341.000000)">
-                      <g id="Group-3-Copy-2" transform="translate(132.000000, 2332.000000)">
-                        <polygon id="Shape-Copy" points="10.6 9 6 13.3513514 1.4 9 0 10.3243243 6 16 12 10.3243243"/>
-                      </g>
-                    </g>
-                  </g>
-                </svg>
-              </div>
-              <div class="custom-carousel__nav-item custom-carousel__right" @click="nextSlide">
-                <svg width="12px" height="7px" viewBox="0 0 12 7" version="1.1" xmlns="http://www.w3.org/2000/svg" class="custom-carousel__right-arrow">
-                  <g id="Page-1" stroke-width="1">
-                    <g transform="translate(-132.000000, -2341.000000)">
-                      <g id="Group-3-Copy-2" transform="translate(132.000000, 2332.000000)">
-                        <polygon id="Shape-Copy" points="10.6 9 6 13.3513514 1.4 9 0 10.3243243 6 16 12 10.3243243"/>
-                      </g>
-                    </g>
-                  </g>
-                </svg>
-              </div>
-              <div class="custom-carousel__paginator"></div>
-            </div>
-          </div>
+            <!--<div class="custom-carousel__nav">-->
+              <!--<div class="custom-carousel__nav-item custom-carousel__left" @click="previousSlide">-->
+                <!--<svg width="12px" height="7px" viewBox="0 0 12 7" version="1.1" xmlns="http://www.w3.org/2000/svg" class="custom-carousel__left-arrow">-->
+                  <!--<g id="Page-1" stroke-width="1">-->
+                    <!--<g transform="translate(-132.000000, -2341.000000)">-->
+                      <!--<g id="Group-3-Copy-2" transform="translate(132.000000, 2332.000000)">-->
+                        <!--<polygon id="Shape-Copy" points="10.6 9 6 13.3513514 1.4 9 0 10.3243243 6 16 12 10.3243243"/>-->
+                      <!--</g>-->
+                    <!--</g>-->
+                  <!--</g>-->
+                <!--</svg>-->
+              <!--</div>-->
+              <!--<div class="custom-carousel__nav-item custom-carousel__right" @click="nextSlide">-->
+                <!--<svg width="12px" height="7px" viewBox="0 0 12 7" version="1.1" xmlns="http://www.w3.org/2000/svg" class="custom-carousel__right-arrow">-->
+                  <!--<g id="Page-1" stroke-width="1">-->
+                    <!--<g transform="translate(-132.000000, -2341.000000)">-->
+                      <!--<g id="Group-3-Copy-2" transform="translate(132.000000, 2332.000000)">-->
+                        <!--<polygon id="Shape-Copy" points="10.6 9 6 13.3513514 1.4 9 0 10.3243243 6 16 12 10.3243243"/>-->
+                      <!--</g>-->
+                    <!--</g>-->
+                  <!--</g>-->
+                <!--</svg>-->
+              <!--</div>-->
+              <!--<div class="custom-carousel__paginator"></div>-->
+            <!--</div>-->
+          </rt-card-carousel>
         </div>
       </div>
     </div>
@@ -441,143 +441,143 @@
   </div>
 </template>
 
-<script>
-  const componentsList = {};
+<!--<script>-->
+  <!--const componentsList = {};-->
 
-  export default {
-    name: "VCOD",
-    components: componentsList,
-    data () {
-      return {
-        mobileLayout: window.innerWidth <= 767,
-        parentArray: [],
-        element: null,
-        parent: null,
-        activeSlideIndex: null,
-        nextSlideIndex: 0
-      }
-    },
-    mounted () {
-      this.element = document.querySelector('.carousel-card');
-      this.parent = this.element.parentNode;
-      for(let i = 0; i < this.parent.children.length; i++) {
-        if (this.parent.children[i].classList.contains('carousel-card')) {
-          this.parentArray.push(this.parent.children[i]);
-        }
-      }
-      window.addEventListener('resize', () => {
-        this.mobileLayout = window.innerWidth <= 767
-      });
-      this.positionCarouselCards();
+  <!--export default {-->
+    <!--name: "VCOD",-->
+    <!--components: componentsList,-->
+    <!--data () {-->
+      <!--return {-->
+        <!--mobileLayout: window.innerWidth <= 767,-->
+        <!--parentArray: [],-->
+        <!--element: null,-->
+        <!--parent: null,-->
+        <!--activeSlideIndex: null,-->
+        <!--nextSlideIndex: 0-->
+      <!--}-->
+    <!--},-->
+    <!--mounted () {-->
+      <!--this.element = document.querySelector('.carousel-card');-->
+      <!--this.parent = this.element.parentNode;-->
+      <!--for(let i = 0; i < this.parent.children.length; i++) {-->
+        <!--if (this.parent.children[i].classList.contains('carousel-card')) {-->
+          <!--this.parentArray.push(this.parent.children[i]);-->
+        <!--}-->
+      <!--}-->
+      <!--window.addEventListener('resize', () => {-->
+        <!--this.mobileLayout = window.innerWidth <= 767-->
+      <!--});-->
+      <!--this.positionCarouselCards();-->
 
-      if(this.mobileLayout) {
-        this.addPaginator();
-        this.fixCardHeightMobile();
-        for(let i = 0; i < this.parentArray.length; i++) {
-          document.querySelector('.custom-carousel__paginator').children[i].addEventListener('click', (e) => {
-            let activePaginatorIndex = this.getChildNumber(e.target);
-            this.mobileCarousel(activePaginatorIndex);
-          })
-        }
-      }
-    },
-    methods: {
-      addPaginator() {
-        for(let i = 0; i < this.parentArray.length; i++) {
-          let paginatorItem = document.createElement('div');
-          paginatorItem.classList.add('custom-carousel__paginator-item');
-          document.querySelector('.custom-carousel__paginator').appendChild(paginatorItem);
-          if(i === this.nextSlideIndex) {
-            document.querySelector('.custom-carousel__paginator-item').classList.add('custom-carousel__paginator-item--active');
-          }
-        }
-      },
+      <!--if(this.mobileLayout) {-->
+        <!--this.addPaginator();-->
+        <!--this.fixCardHeightMobile();-->
+        <!--for(let i = 0; i < this.parentArray.length; i++) {-->
+          <!--document.querySelector('.custom-carousel__paginator').children[i].addEventListener('click', (e) => {-->
+            <!--let activePaginatorIndex = this.getChildNumber(e.target);-->
+            <!--this.mobileCarousel(activePaginatorIndex);-->
+          <!--})-->
+        <!--}-->
+      <!--}-->
+    <!--},-->
+    <!--methods: {-->
+      <!--addPaginator() {-->
+        <!--for(let i = 0; i < this.parentArray.length; i++) {-->
+          <!--let paginatorItem = document.createElement('div');-->
+          <!--paginatorItem.classList.add('custom-carousel__paginator-item');-->
+          <!--document.querySelector('.custom-carousel__paginator').appendChild(paginatorItem);-->
+          <!--if(i === this.nextSlideIndex) {-->
+            <!--document.querySelector('.custom-carousel__paginator-item').classList.add('custom-carousel__paginator-item&#45;&#45;active');-->
+          <!--}-->
+        <!--}-->
+      <!--},-->
 
-      getChildNumber(node) {
-        return Array.prototype.indexOf.call(node.parentNode.childNodes, node);
-      },
+      <!--getChildNumber(node) {-->
+        <!--return Array.prototype.indexOf.call(node.parentNode.childNodes, node);-->
+      <!--},-->
 
-      mobileCarousel(index) {
-        let carouselPaginatorWrapper = document.querySelector('.custom-carousel__paginator');
-        let slideWidth = document.querySelector('.carousel-card').offsetWidth;
-        let carouselWrapperPadding = +window.getComputedStyle(document.querySelector('.custom-carousel')).paddingLeft.replace('px', '')
-        for(let i = 0; i < this.parentArray.length; i++) {
-          carouselPaginatorWrapper.children[i].classList.remove('custom-carousel__paginator-item--active');
-        }
-        carouselPaginatorWrapper.children[index].classList.add('custom-carousel__paginator-item--active');
-        document.querySelector('.custom-carousel').scrollBy({
-          top: 0,
-          left: (slideWidth * index + (20 * index) - ((window.innerWidth - slideWidth) / 2) + carouselWrapperPadding) - document.querySelector('.custom-carousel').scrollLeft,
-          behavior: 'smooth'
-        });
-//        document.querySelector('.custom-carousel').scrollBy(slideWidth * index + (20 * index) - ((window.innerWidth - slideWidth) / 2) + carouselWrapperPadding);
-      },
+      <!--mobileCarousel(index) {-->
+        <!--let carouselPaginatorWrapper = document.querySelector('.custom-carousel__paginator');-->
+        <!--let slideWidth = document.querySelector('.carousel-card').offsetWidth;-->
+        <!--let carouselWrapperPadding = +window.getComputedStyle(document.querySelector('.custom-carousel')).paddingLeft.replace('px', '')-->
+        <!--for(let i = 0; i < this.parentArray.length; i++) {-->
+          <!--carouselPaginatorWrapper.children[i].classList.remove('custom-carousel__paginator-item&#45;&#45;active');-->
+        <!--}-->
+        <!--carouselPaginatorWrapper.children[index].classList.add('custom-carousel__paginator-item&#45;&#45;active');-->
+        <!--document.querySelector('.custom-carousel').scrollBy({-->
+          <!--top: 0,-->
+          <!--left: (slideWidth * index + (20 * index) - ((window.innerWidth - slideWidth) / 2) + carouselWrapperPadding) - document.querySelector('.custom-carousel').scrollLeft,-->
+          <!--behavior: 'smooth'-->
+        <!--});-->
+<!--//        document.querySelector('.custom-carousel').scrollBy(slideWidth * index + (20 * index) - ((window.innerWidth - slideWidth) / 2) + carouselWrapperPadding);-->
+      <!--},-->
 
-      positionCarouselCards() {
-        if(!this.mobileLayout){
-          this.parentArray.forEach((el, i)=>{
-            if(((i + 1) < this.nextSlideIndex) || (i - 1) > this.nextSlideIndex) {
-              el.classList.remove('carousel-card--next', 'carousel-card--previous');
-              if(((i - 2) === this.nextSlideIndex)) {
-                el.classList.add('carousel-card--next-via-one');
-              } else if(((i + 2) === this.nextSlideIndex)) {
-                el.classList.add('carousel-card--previous-via-one');
-              }
-            }
-            if((i + 1) === this.nextSlideIndex || (this.nextSlideIndex === 0 && i === this.parentArray.length - 1)){
-              el.classList.remove('carousel-card--next-via-one', 'carousel-card--previous-via-one');
-              el.classList.add('carousel-card--previous');
-            }
-            if((i - 1) === this.nextSlideIndex || (this.nextSlideIndex === this.parentArray.length - 1 && i === 0)){
-              el.classList.remove('carousel-card--next-via-one', 'carousel-card--previous-via-one');
-              el.classList.add('carousel-card--next');
-            }
-          });
-        }
-      },
+      <!--positionCarouselCards() {-->
+        <!--if(!this.mobileLayout){-->
+          <!--this.parentArray.forEach((el, i)=>{-->
+            <!--if(((i + 1) < this.nextSlideIndex) || (i - 1) > this.nextSlideIndex) {-->
+              <!--el.classList.remove('carousel-card&#45;&#45;next', 'carousel-card&#45;&#45;previous');-->
+              <!--if(((i - 2) === this.nextSlideIndex)) {-->
+                <!--el.classList.add('carousel-card&#45;&#45;next-via-one');-->
+              <!--} else if(((i + 2) === this.nextSlideIndex)) {-->
+                <!--el.classList.add('carousel-card&#45;&#45;previous-via-one');-->
+              <!--}-->
+            <!--}-->
+            <!--if((i + 1) === this.nextSlideIndex || (this.nextSlideIndex === 0 && i === this.parentArray.length - 1)){-->
+              <!--el.classList.remove('carousel-card&#45;&#45;next-via-one', 'carousel-card&#45;&#45;previous-via-one');-->
+              <!--el.classList.add('carousel-card&#45;&#45;previous');-->
+            <!--}-->
+            <!--if((i - 1) === this.nextSlideIndex || (this.nextSlideIndex === this.parentArray.length - 1 && i === 0)){-->
+              <!--el.classList.remove('carousel-card&#45;&#45;next-via-one', 'carousel-card&#45;&#45;previous-via-one');-->
+              <!--el.classList.add('carousel-card&#45;&#45;next');-->
+            <!--}-->
+          <!--});-->
+        <!--}-->
+      <!--},-->
 
-      nextSlide() {
-        if(!this.mobileLayout) {
-          for(let i = 0; i < this.parentArray.length; i++){
-            if(this.parent.children[i].classList.contains('carousel-card--active')){
-              this.activeSlideIndex = this.parentArray.indexOf(this.parent.children[i]);
-              this.nextSlideIndex = (this.activeSlideIndex +  1) % 8 ? this.activeSlideIndex + 1 : 0;
-            }
-          }
-          this.parentArray[this.activeSlideIndex].classList.remove('carousel-card--active');
-          this.parentArray[this.nextSlideIndex].classList.add('carousel-card--active');
-          this.parentArray[this.nextSlideIndex].classList
-            .remove('carousel-card--next', 'carousel-card--previous', 'carousel-card--next-via-one', 'carousel-card--previous-via-one');
-          this.positionCarouselCards()
-        }
-      },
+      <!--nextSlide() {-->
+        <!--if(!this.mobileLayout) {-->
+          <!--for(let i = 0; i < this.parentArray.length; i++){-->
+            <!--if(this.parent.children[i].classList.contains('carousel-card&#45;&#45;active')){-->
+              <!--this.activeSlideIndex = this.parentArray.indexOf(this.parent.children[i]);-->
+              <!--this.nextSlideIndex = (this.activeSlideIndex +  1) % 8 ? this.activeSlideIndex + 1 : 0;-->
+            <!--}-->
+          <!--}-->
+          <!--this.parentArray[this.activeSlideIndex].classList.remove('carousel-card&#45;&#45;active');-->
+          <!--this.parentArray[this.nextSlideIndex].classList.add('carousel-card&#45;&#45;active');-->
+          <!--this.parentArray[this.nextSlideIndex].classList-->
+            <!--.remove('carousel-card&#45;&#45;next', 'carousel-card&#45;&#45;previous', 'carousel-card&#45;&#45;next-via-one', 'carousel-card&#45;&#45;previous-via-one');-->
+          <!--this.positionCarouselCards()-->
+        <!--}-->
+      <!--},-->
 
-      previousSlide() {
-        if(!this.mobileLayout) {
-          for(let i = 0; i < this.parentArray.length; i++){
-            if(this.parent.children[i].classList.contains('carousel-card--active')){
-              this.activeSlideIndex = this.parentArray.indexOf(this.parent.children[i]);
-              this.nextSlideIndex = (this.activeSlideIndex -  1) >= 0 ? this.activeSlideIndex - 1 : this.parentArray.length - 1;
-            }
-          }
-          this.parentArray[this.activeSlideIndex].classList.remove('carousel-card--active');
-          this.parentArray[this.nextSlideIndex].classList.add('carousel-card--active');
-          this.parentArray[this.nextSlideIndex].classList
-            .remove('carousel-card--next', 'carousel-card--previous', 'carousel-card--next-via-one', 'carousel-card--previous-via-one');
-          this.positionCarouselCards()
-        }
-      },
+      <!--previousSlide() {-->
+        <!--if(!this.mobileLayout) {-->
+          <!--for(let i = 0; i < this.parentArray.length; i++){-->
+            <!--if(this.parent.children[i].classList.contains('carousel-card&#45;&#45;active')){-->
+              <!--this.activeSlideIndex = this.parentArray.indexOf(this.parent.children[i]);-->
+              <!--this.nextSlideIndex = (this.activeSlideIndex -  1) >= 0 ? this.activeSlideIndex - 1 : this.parentArray.length - 1;-->
+            <!--}-->
+          <!--}-->
+          <!--this.parentArray[this.activeSlideIndex].classList.remove('carousel-card&#45;&#45;active');-->
+          <!--this.parentArray[this.nextSlideIndex].classList.add('carousel-card&#45;&#45;active');-->
+          <!--this.parentArray[this.nextSlideIndex].classList-->
+            <!--.remove('carousel-card&#45;&#45;next', 'carousel-card&#45;&#45;previous', 'carousel-card&#45;&#45;next-via-one', 'carousel-card&#45;&#45;previous-via-one');-->
+          <!--this.positionCarouselCards()-->
+        <!--}-->
+      <!--},-->
 
-      fixCardHeightMobile() {
-        let maxHeight = 0;
-        for(let i = 0; i < this.parentArray.length; i++) {
-          let cardHeight = +window.getComputedStyle(document.querySelector('.custom-carousel').children[i]).height.replace('px', '');
-          maxHeight = maxHeight < cardHeight ? cardHeight : maxHeight;
-        }
-        for(let i = 0; i < this.parentArray.length; i++) {
-          document.querySelector('.custom-carousel').children[i].style.height = maxHeight + 'px';
-        }
-      }
-    }
-  }
-</script>
+      <!--fixCardHeightMobile() {-->
+        <!--let maxHeight = 0;-->
+        <!--for(let i = 0; i < this.parentArray.length; i++) {-->
+          <!--let cardHeight = +window.getComputedStyle(document.querySelector('.custom-carousel').children[i]).height.replace('px', '');-->
+          <!--maxHeight = maxHeight < cardHeight ? cardHeight : maxHeight;-->
+        <!--}-->
+        <!--for(let i = 0; i < this.parentArray.length; i++) {-->
+          <!--document.querySelector('.custom-carousel').children[i].style.height = maxHeight + 'px';-->
+        <!--}-->
+      <!--}-->
+    <!--}-->
+  <!--}-->
+<!--</script>-->
