+++ src/lib/css/components/swiper.styl
@@ -35,6 +35,7 @@ scrollHSpacePC = 70px
     //width: calc(100vw - 140px)
     margin: 0 auto
     width calc(100vw - (100vw - 100%))
+    max-width 100vw
     @media (max-width: tablet-upper-limit)
       padding-right 0
       padding-left 0
@@ -64,10 +65,13 @@ scrollHSpacePC = 70px
     user-select none
     @media (max-width: laptop-upper-limit) and (min-width: desktop-lower-limit)
       max-width calc(275px + 20px)
+      flex-basis calc(275px + 20px)
     @media (max-width: tablet-upper-limit)
       width calc(35vw + 20px)
+      flex-basis calc(35vw + 20px)
     @media (max-width: mobile-upper-limit)
       width calc(75vw + 20px)
+      flex-basis calc(75vw + 20px)
 
     &.marked
       transform scale(1.02)
